refactor: centralize OutboxRecoveryService lifecycle in ServicesProvider
Move OutboxRecoveryService initialization from useWallet.ts to ServicesProvider.tsx
to prevent the repeated start/stop cycles (~22 times) that occurred when multiple
components using useWallet() each had their own lifecycle management.

Changes:
- ServicesProvider.tsx: Add centralized OutboxRecoveryService initialization
  after Nostr connects, with proper cleanup on wallet-loaded and unmount
- useWallet.ts: Remove duplicate lifecycle management useEffect (38 lines)
- OutboxRecoveryService.ts: Add guard to skip redundant starts when already
  running for the same address

This ensures the recovery service starts exactly once during app initialization
regardless of how many components use the useWallet() hook.
